Rename component filenames to match component names
