+++ app/__tests__/popup.test.jsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { shallow } from "enzyme";
-import { Popup } from "./../containers/popup";
+import { Popup } from "./../containers/Popup";
 
 import * as MessageService from "../message_service";
